+++ src/main/index.ts
@@ -6,8 +6,8 @@ import { authPackage } from './auth';
 import { setupCommentIndexer } from './commentIndexer';
 import { setupIndex } from './indexer';
 import { setupLSPs } from './lsp';
-import setupMainMenu from './mainMenu';
-import mainWindow from './mainWindow';
+import setupMainMenu from './menu';
+import mainWindow from './window';
 import { setupSearch } from './search';
 import setupApplicationsFolder from './setup/appFolder';
 import setupAuth from './setup/auth';
